requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==4.2
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
